По замечаниям. Перенесет state city в редьюсер data
