+++ src/reducer/app/app.js
@@ -2,22 +2,16 @@ import {extend} from "../../utils.js";
 import {SortTypes} from "../../const.js";
 
 const initialState = {
-  city: ``,
   sortType: SortTypes.POPULAR,
   activeOfferId: -1
 };
 
 const ActionType = {
-  CHANGE_CITY: `CHANGE_CITY`,
   CHANGE_SORT: `CHANGE_SORT`,
   CHANGE_ACTIVE_OFFER_ID: `CHANGE_ACTIVE_OFFER_ID`,
 };
 
 const ActionCreator = {
-  changeCity: (city) => ({
-    type: ActionType.CHANGE_CITY,
-    payload: city
-  }),
   changeSort: (sortType) => ({
     type: ActionType.CHANGE_SORT,
     payload: sortType
@@ -30,8 +24,6 @@ const ActionCreator = {
 
 const reducer = (state = initialState, action) => {
   switch (action.type) {
-    case ActionType.CHANGE_CITY:
-      return extend(state, {city: action.payload});
     case ActionType.CHANGE_SORT:
       return extend(state, {sortType: action.payload});
     case ActionType.CHANGE_ACTIVE_OFFER_ID:
