+++ src/reducer/app/app.test.js
@@ -3,37 +3,19 @@ import {reducer, ActionType, ActionCreator} from "./app.js";
 
 it(`Reducer without additional parameters should return initial state`, () => {
   expect(reducer(void 0, {})).toEqual({
-    city: ``,
     sortType: SortTypes.POPULAR,
     activeOfferId: -1
   });
 });
 
-it(`Reducer should change city name by a given value`, () => {
-  expect(reducer({
-    city: ``,
-    sortType: SortTypes.POPULAR,
-    activeOfferId: -1,
-  }, {
-    type: ActionType.CHANGE_CITY,
-    payload: `Paris`,
-  })).toEqual({
-    city: `Paris`,
-    sortType: SortTypes.POPULAR,
-    activeOfferId: -1,
-  });
-});
-
 it(`Reducer should change sort type by a given value`, () => {
   expect(reducer({
-    city: ``,
     sortType: SortTypes.POPULAR,
     activeOfferId: -1,
   }, {
     type: ActionType.CHANGE_SORT,
     payload: SortTypes.PRICE_HIGH_LOW,
   })).toEqual({
-    city: ``,
     activeOfferId: -1,
     sortType: SortTypes.PRICE_HIGH_LOW,
   });
@@ -41,14 +23,12 @@ it(`Reducer should change sort type by a given value`, () => {
 
 it(`Reducer should change active offer id by a given value`, () => {
   expect(reducer({
-    city: ``,
     sortType: SortTypes.POPULAR,
     activeOfferId: -1
   }, {
     type: ActionType.CHANGE_ACTIVE_OFFER_ID,
     payload: 2,
   })).toEqual({
-    city: ``,
     sortType: SortTypes.POPULAR,
     activeOfferId: 2,
   });
@@ -56,13 +36,6 @@ it(`Reducer should change active offer id by a given value`, () => {
 
 
 describe(`Action creators work correctly`, () => {
-  it(`Action creator for changing city returns correct action`, () => {
-    expect(ActionCreator.changeCity(`Paris`)).toEqual({
-      type: ActionType.CHANGE_CITY,
-      payload: `Paris`,
-    });
-  });
-
   it(`Action creator for changing sort returns correct action`, () => {
     expect(ActionCreator.changeSort(SortTypes.TOP_RATED_FIRST)).toEqual({
       type: ActionType.CHANGE_SORT,
