+++ src/reducer/app/selectors.js
@@ -2,10 +2,6 @@ import NameSpace from "../name-space.js";
 
 const NAME_SPACE = NameSpace.APP;
 
-const getCity = (state) => {
-  return state[NAME_SPACE].city;
-};
-
 const getSortType = (state) => {
   return state[NAME_SPACE].sortType;
 };
@@ -14,4 +10,4 @@ const getActiveOfferId = (state) => {
   return state[NAME_SPACE].activeOfferId;
 };
 
-export {getCity, getSortType, getActiveOfferId};
+export {getSortType, getActiveOfferId};
