+++ src/reducer/data/data.js
@@ -2,17 +2,23 @@ import {extend} from "../../utils.js";
 import offerAdapter from "../../adapters/offer.js";
 
 const initialState = {
+  city: ``,
   offers: [],
 };
 
 const ActionType = {
   LOAD_OFFERS: `LOAD_OFFERS`,
+  CHANGE_CITY: `CHANGE_CITY`,
 };
 
 const ActionCreator = {
   loadOffers: (loadedOffers) => ({
     type: ActionType.LOAD_OFFERS,
     payload: loadedOffers
+  }),
+  changeCity: (city) => ({
+    type: ActionType.CHANGE_CITY,
+    payload: city
   }),
 };
 
@@ -22,6 +28,7 @@ const Operation = {
       .then((response) => {
         const loadedOffers = response.data.map((offer) => offerAdapter(offer));
         dispatch(ActionCreator.loadOffers(loadedOffers));
+        dispatch(ActionCreator.changeCity(loadedOffers[0].city.name));
         return loadedOffers;
       });
   }
@@ -31,6 +38,8 @@ const reducer = (state = initialState, action) => {
   switch (action.type) {
     case ActionType.LOAD_OFFERS:
       return extend(state, {offers: action.payload});
+    case ActionType.CHANGE_CITY:
+      return extend(state, {city: action.payload});
   }
 
   return state;
